+++ test/commands/mdapi/retrieve.test.ts
@@ -59,6 +59,7 @@ describe('force:mdapi:retrieve', () => {
   }
 
   const runRetrieveCmd = async (params: string[]) => {
+    // @ts-expect-error type mismatch between oclif/core v1 and v2
     const cmd = new TestRetrieve(params, oclifConfigStub);
     stubMethod(sandbox, cmd, 'assignOrg').callsFake(() => {
       const orgStub = fromStub(