+++ test/commands/source/delete.test.ts
@@ -57,6 +57,7 @@ describe('force:source:delete', () => {
   }
 
   const runDeleteCmd = async (params: string[]) => {
+    // @ts-expect-error type mismatch between oclif/core v1 and v2
     const cmd = new TestDelete(params, oclifConfigStub);
     stubMethod(sandbox, SfProject, 'resolveProjectPath').resolves(join('path', 'to', 'package'));
     stubMethod(sandbox, cmd, 'assignProject').callsFake(() => {