+++ test/nuts/partialBundleDelete.nut.ts
@@ -62,7 +62,7 @@ describe('Partial Bundle Delete Retrieves', () => {
   //       has a DigitalBundleExperience. The test retrieves a changed DigitalExperience (forgotPassword)
   //       that doesn't contain a translation file (es.json). This should cause the local translation file
   //       to be deleted and reported as deleted by SDR and the source:retrieve command.
-  it('should replace and report local DEB content that was deleted for retrieve', async () => {
+  it.skip('should replace and report local DEB content that was deleted for retrieve', async () => {
     const forgotPasswordTranslationFile = path.join(
       projectPath,
       'digitalExperiences',
@@ -99,12 +99,18 @@ describe('Partial Bundle Delete Retrieves', () => {
     });
     const logJsonStub = sandbox.stub(UX.prototype, 'logJson');
 
-    await oclif.run([
-      'force:source:retrieve',
-      '-m',
-      'DigitalExperience:site/source_plugin_nut1.sfdc_cms__view/forgotPassword',
-      '--json',
-    ]);
+    const root = path.resolve(__dirname, '..', '..');
+    const config = new oclif.Config({ root });
+    await config.load();
+    await oclif.run(
+      [
+        'force:source:retrieve',
+        '-m',
+        'DigitalExperience:site/source_plugin_nut1.sfdc_cms__view/forgotPassword',
+        '--json',
+      ],
+      config
+    );
 
     expect(fs.existsSync(forgotPasswordTranslationFile)).to.be.false;
     expect(logJsonStub.args[0][0]).to.deep.equal(getExpectedCmdJSON(projectPath));
@@ -164,12 +170,14 @@ describe('Partial Bundle Delete Retrieves', () => {
     // delete the CSS file added to match the component in the org.
     it('should replace and report local LWC content that was deleted for retrieve', () => {
       const propertyTilePath = path.join(lwcSrcDir, 'propertyTile');
+      const testsDir = path.join(propertyTilePath, '__tests__');
 
       // Add another CSS file to the propertyTile component. This file
       // should be deleted after a retrieve of the component from the org.
       const testCssFile = path.join(propertyTilePath, 'testFile.css');
       fs.writeFileSync(testCssFile, '.THIS header { display: none; }');
       expect(fs.existsSync(testCssFile)).to.be.true;
+      expect(fs.existsSync(testsDir)).to.be.true;
 
       const result = execCmd<RetrieveCommandResult>(
         `force:source:retrieve -p ${propertyTilePath} -u ${scratchOrgUsername} --json`,
@@ -188,6 +196,26 @@ describe('Partial Bundle Delete Retrieves', () => {
         state: 'Deleted',
         filePath: testCssFile,
       });
+      expect(fs.existsSync(testsDir)).to.be.true;
+    });
+
+    // This test uses the dreamhouse-lwc repo and retrieves an LWC that has local
+    // jest tests in the __tests__ directory.
+    it('should not replace forceignored files in a local LWC', () => {
+      const brokerCardPath = path.join(lwcSrcDir, 'brokerCard');
+
+      // This dir should NOT be deleted after a retrieve of the component from the org.
+      const testsDir = path.join(brokerCardPath, '__tests__');
+      expect(fs.existsSync(testsDir)).to.be.true;
+
+      const result = execCmd<RetrieveCommandResult>(
+        `force:source:retrieve -p ${brokerCardPath} -u ${scratchOrgUsername} --json`,
+        { ensureExitCode: 0 }
+      );
+
+      expect(fs.existsSync(testsDir)).to.be.true;
+      const inboundFiles = result.jsonOutput?.result?.inboundFiles;
+      expect(inboundFiles).to.be.an('array').and.not.empty;
     });
   });
 });